Move everything servlet-related to ninja.servlet
