+++ ninja-core/src/main/java/ninja/servlet/ResponseStreamsServlet.java
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package ninja.utils;
+package ninja.servlet;
+
+import ninja.utils.ResponseStreams;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -24,9 +26,9 @@ import javax.servlet.http.HttpServletResponse;
 
 /**
  * Make sure to only write to either the OutputStream OR the Writer...
- * 
+ *
  * @author rbauer
- * 
+ *
  */
 public class ResponseStreamsServlet implements ResponseStreams {
 
@@ -39,9 +41,9 @@ public class ResponseStreamsServlet implements ResponseStreams {
 
     /**
      * Get the output stream to write the response.
-     * 
+     *
      * Must not be used if getWriter has been called.
-     * 
+     *
      * @return The output stream
      */
     public OutputStream getOutputStream() throws IOException {
@@ -50,9 +52,9 @@ public class ResponseStreamsServlet implements ResponseStreams {
 
     /**
      * Get the writer to write the response.
-     * 
+     *
      * Must not be used if getOutputStream has been called.
-     * 
+     *
      * @return The writer
      */
     public Writer getWriter() throws IOException {
