+++ ninja-core/src/main/java/ninja/servlet/async/Servlet3AsyncStrategy.java
@@ -14,38 +14,40 @@
  * limitations under the License.
  */
 
-package ninja.async;
+package ninja.servlet.async;
 
 import ninja.Context;
 import ninja.Result;
+import ninja.utils.ResultHandler;
 
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicReference;
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * @author James Roper
  */
-public class BlockingAsyncStrategy implements AsyncStrategy {
-    private final CountDownLatch requestCompleteLatch = new CountDownLatch(1);
-    private final AtomicReference<Result> result = new AtomicReference<Result>();
+public class Servlet3AsyncStrategy implements AsyncStrategy {
+    private final ResultHandler resultHandler;
+    private final HttpServletRequest request;
+
+    public Servlet3AsyncStrategy(ResultHandler resultHandler,
+                                 HttpServletRequest request) {
+        this.resultHandler = resultHandler;
+        this.request = request;
+    }
 
     @Override
     public void handleAsync() {
+        request.startAsync();
     }
 
     @Override
     public Result controllerReturned() {
-        try {
-            requestCompleteLatch.await();
-            return this.result.get();
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+        return null;
     }
 
     @Override
     public void returnResultAsync(Result result, Context context) {
-        this.result.set(result);
-        requestCompleteLatch.countDown();
+        resultHandler.handleResult(result, context);
+        request.getAsyncContext().complete();
     }
 }
