+++ ninja-core/src/test/java/ninja/servlet/ContextImplTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ninja;
+package ninja.servlet;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.nullValue;
@@ -33,9 +33,9 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import ninja.*;
 import ninja.bodyparser.BodyParserEngine;
 import ninja.bodyparser.BodyParserEngineManager;
-import ninja.servlet.ContextImpl;
 import ninja.session.FlashCookie;
 import ninja.session.SessionCookie;
 
@@ -54,22 +54,22 @@ import com.google.common.collect.Maps;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ContextImplTest {
-    
+
     @Mock
     private SessionCookie sessionCookie;
-    
+
     @Mock
     private FlashCookie flashCookie;
-    
+
     @Mock
-    private BodyParserEngineManager bodyParserEngineManager;    
-    
+    private BodyParserEngineManager bodyParserEngineManager;
+
     @Mock
     private HttpServletRequest httpServletRequest;
-    
+
     @Mock
     private HttpServletResponse httpServletResponse;
-    
+
     @Mock
     private Route route;
 
@@ -81,7 +81,7 @@ public class ContextImplTest {
 
     @Mock
     private BodyParserEngine bodyParserEngine;
-    
+
     private ContextImpl context;
 
     @Before
@@ -90,21 +90,21 @@ public class ContextImplTest {
         //According to servlet spec the following will be returned:
         when(httpServletRequest.getContextPath()).thenReturn("");
         when(httpServletRequest.getRequestURI()).thenReturn("/");
-        
-        
+
+
         context = new ContextImpl(bodyParserEngineManager, flashCookie, sessionCookie,
                 resultHandler, validation);
     }
 
     @Test
     public void testGetRequestUri() {
-        
+
         //say the httpServletRequest to return a certain value:
         when(httpServletRequest.getRequestURI()).thenReturn("/index");
-        
+
         //init the context from a (mocked) servlet
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //make sure this is correct
         assertEquals("/index", context.getRequestUri());
     }
@@ -118,7 +118,7 @@ public class ContextImplTest {
         //generate an arbitrary result:
         Result result = Results.html();
         result.addCookie(cookie);
-        
+
         //finalize the headers => the cookies must be copied over to the servletcookies
         context.finalizeHeaders(result);
 
@@ -133,35 +133,35 @@ public class ContextImplTest {
         assertThat(resultCookie.getSecure(), equalTo(false));
         assertThat(resultCookie.getMaxAge(), equalTo(-1));
     }
-    
+
     @Test
     public void getCookieTest() {
         javax.servlet.http.Cookie servletCookie1 = new javax.servlet.http.Cookie("contextCookie1", "theValue1");
         javax.servlet.http.Cookie servletCookie2 = new javax.servlet.http.Cookie("contextCookie2", "theValue2");
         javax.servlet.http.Cookie [] servletCookies = {servletCookie1, servletCookie2};
-        
+
         when(httpServletRequest.getCookies()).thenReturn(servletCookies);
-        
+
         context.init(httpServletRequest, httpServletResponse);
-        
+
         // negative test:
         ninja.Cookie doesNotExist = context.getCookie("doesNotExist");
         assertNull(doesNotExist);
-        
+
         // test  against cookie that is really there
         ninja.Cookie cookie1 = context.getCookie("contextCookie1");
-        
+
         assertEquals(cookie1.getName(), "contextCookie1");
         assertEquals(cookie1.getValue(), "theValue1");
-        
+
         // test 2 against cookie that is really there
         ninja.Cookie cookie2= context.getCookie("contextCookie2");
-        
+
         assertEquals(cookie2.getName(), "contextCookie2");
         assertEquals(cookie2.getValue(), "theValue2");
     }
 
-        
+
     @Test
     public void hasCookieTest() {
         javax.servlet.http.Cookie servletCookie1 = new javax.servlet.http.Cookie("contextCookie1", "theValue1");
@@ -184,202 +184,202 @@ public class ContextImplTest {
 
     @Test
     public void getCookiesTest() {
-        
+
         javax.servlet.http.Cookie servletCookie1 = new javax.servlet.http.Cookie("contextCookie1", "theValue");
         javax.servlet.http.Cookie servletCookie2 = new javax.servlet.http.Cookie("contextCookie2", "theValue");
         javax.servlet.http.Cookie [] servletCookiesEmpty = {};
         javax.servlet.http.Cookie [] servletCookies = {servletCookie1, servletCookie2};
-        
+
         when(httpServletRequest.getCookies()).thenReturn(servletCookiesEmpty);
-        
+
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //test when there are no cookies.
         assertEquals(0, context.getCookies().size());
-        
+
         // now return some cookies:
         when(httpServletRequest.getCookies()).thenReturn(servletCookies);
-        
+
 
         assertEquals(2, context.getCookies().size());
 
 
     }
-    
-    
+
+
     @Test
     public void testGetPathParameter() {
     	//init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //mock a parametermap:
         Map<String, String> parameterMap = Maps.newHashMap();
         parameterMap.put("parameter", "parameter");
-        
+
         //and return the parameter map when any parameter is called...
-        when(route.getPathParametersEncoded(Matchers.anyString())).thenReturn(parameterMap);	
-		
+        when(route.getPathParametersEncoded(Matchers.anyString())).thenReturn(parameterMap);
+
         context.setRoute(route);
-        
+
         //this parameter is not there and must return null
-        assertEquals(null, context.getPathParameter("parameter_not_set")); 
-        
-        assertEquals("parameter", context.getPathParameter("parameter")); 
+        assertEquals(null, context.getPathParameter("parameter_not_set"));
+
+        assertEquals("parameter", context.getPathParameter("parameter"));
 
     }
-    
+
     @Test
     public void testGetPathParameterDecodingWorks() {
         //init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //mock a parametermap:
         Map<String, String> parameterMap = Maps.newHashMap();
         parameterMap.put("parameter", "blue%2Fred%3Fand+green%E2%82%AC%2f");
-        
+
         //and return the parameter map when any parameter is called...
-        when(route.getPathParametersEncoded(Matchers.anyString())).thenReturn(parameterMap);   
-        
+        when(route.getPathParametersEncoded(Matchers.anyString())).thenReturn(parameterMap);
+
         context.setRoute(route);
-        
+
         //that is how the above parameter looks decoded correctly:
-        assertEquals("blue/red?and+green€/", context.getPathParameter("parameter")); 
+        assertEquals("blue/red?and+green€/", context.getPathParameter("parameter"));
 
     }
-    
-    
+
+
     @Test
     public void testGetPathParameterAsInteger() {
     	//init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //mock a parametermap:
         Map<String, String> parameterMap = Maps.newHashMap();
         parameterMap.put("parameter", "parameter");
-        
+
         //and return the parameter map when any parameter is called...
-        when(route.getPathParametersEncoded(Matchers.anyString())).thenReturn(parameterMap);	
-		
+        when(route.getPathParametersEncoded(Matchers.anyString())).thenReturn(parameterMap);
+
         context.setRoute(route);
-        
+
         //this will not work and return null
-        assertEquals(null, context.getPathParameterAsInteger("parameter")); 
-        
+        assertEquals(null, context.getPathParameterAsInteger("parameter"));
+
         //now set an integer into the parametermap:
         parameterMap.put("parameter", "1");
-        
+
         //this will work and return 1
-        assertEquals(new Integer(1), context.getPathParameterAsInteger("parameter")); 
-        
+        assertEquals(new Integer(1), context.getPathParameterAsInteger("parameter"));
+
     }
-    
+
     @Test
     public void testGetParameter() {
-    	
+
     	//init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //and return the parameter map when any parameter is called...
-        when(httpServletRequest.getParameter("key")).thenReturn("value");	
+        when(httpServletRequest.getParameter("key")).thenReturn("value");
 
         //this will not work and return null
-        assertEquals(null, context.getParameter("key_not_there")); 
-        
+        assertEquals(null, context.getParameter("key_not_there"));
+
         //this will return the default value:
-        assertEquals("defaultValue", context.getParameter("key_not_there", "defaultValue")); 
-        
+        assertEquals("defaultValue", context.getParameter("key_not_there", "defaultValue"));
+
         //this will work as the value is there...
-        assertEquals("value", context.getParameter("key")); 
+        assertEquals("value", context.getParameter("key"));
 
     }
-    
+
     @Test
     public void testGetParameterAsInteger() {
-    	
+
     	//init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //and return the parameter map when any parameter is called...
-        when(httpServletRequest.getParameter("key")).thenReturn("1");	
+        when(httpServletRequest.getParameter("key")).thenReturn("1");
 
         //this will not work and return null
-        assertEquals(null, context.getParameterAsInteger("key_not_there")); 
-        
+        assertEquals(null, context.getParameterAsInteger("key_not_there"));
+
         //this will return the default value:
-        assertEquals(new Integer(100), context.getParameterAsInteger("key_not_there", 100)); 
-        
+        assertEquals(new Integer(100), context.getParameterAsInteger("key_not_there", 100));
+
         //this will work as the value is there...
-        assertEquals(new Integer(1), context.getParameterAsInteger("key"));  
-    	
+        assertEquals(new Integer(1), context.getParameterAsInteger("key"));
+
     }
-    
+
     @Test
     public void testContentTypeGetsConvertedProperlyUponFinalize() {
-        
+
         //init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //this must be Content-Type: application/json; encoding=utf-8
         Result result = Results.json();
-        
+
         context.finalizeHeaders(result);
-        
+
         verify(httpServletResponse).setCharacterEncoding(result.getCharset());
         verify(httpServletResponse).setContentType(result.getContentType());
     }
-    
-    
+
+
     @Test
     public void testContentTypeWithNullEncodingGetsConvertedProperlyUponFinalize() {
-        
+
         //init the context
         context.init(httpServletRequest, httpServletResponse);
-        
+
         //this must be Content-Type: application/json; encoding=utf-8
         Result result = Results.json();
         //force a characterset that is not there. Stupid but tests that its working.
         result.charset(null);
-        
+
         context.finalizeHeaders(result);
-        
+
         //make sure utf-8 is used under all circumstances:
         verify(httpServletResponse).setCharacterEncoding(NinjaConstant.UTF_8);
     }
-    
-    
+
+
     @Test
     public void testGetRequestPathWorksAsExpectedWithContext() {
-        
+
         // we got a context
         when(httpServletRequest.getContextPath()).thenReturn("/my/funky/prefix");
-        
+
         // we got a request uri
-        when(httpServletRequest.getRequestURI()).thenReturn("/my/funky/prefix/myapp/is/here");       
-        
-        
+        when(httpServletRequest.getRequestURI()).thenReturn("/my/funky/prefix/myapp/is/here");
+
+
         context.init(httpServletRequest, httpServletResponse);
-        
+
 
         assertEquals("/myapp/is/here", context.getRequestPath());
-        
+
     }
-    
+
     @Test
     public void testGetRequestPathWorksAsExpectedWithOutContext() {
-        
+
         // we got not context.
         // according to spec it will return an empty string
         when(httpServletRequest.getContextPath()).thenReturn("");
         when(httpServletRequest.getRequestURI()).thenReturn("/index");
         // we got a request uri
-        when(httpServletRequest.getRequestURI()).thenReturn("/myapp/is/here");       
-        
-        
+        when(httpServletRequest.getRequestURI()).thenReturn("/myapp/is/here");
+
+
         context.init(httpServletRequest, httpServletResponse);
-        
+
 
         assertEquals("/myapp/is/here", context.getRequestPath());
-        
+
     }
 
     @Test
@@ -500,163 +500,163 @@ public class ContextImplTest {
 
         assertEquals(charset, context.getAcceptCharset());
     }
-    
+
     /**
      * This is the default mode.
-     * 
+     *
      * We get a Content-Type: application/json and want to parse the incoming json.
      */
     @Test
     public void testParseBodyJsonWorks() {
 
-        
+
         when(httpServletRequest.getContentType()).thenReturn("application/json; charset=utf-8");
-        
+
         //init the context from a (mocked) servlet
         context.init(httpServletRequest, httpServletResponse);
-        
+
         when(bodyParserEngineManager.getBodyParserEngineForContentType("application/json")).thenReturn(bodyParserEngine);
         when(bodyParserEngine.invoke(context, Dummy.class)).thenReturn(new Dummy());
-        
+
         Object o = context.parseBody(Dummy.class);
-        
-        verify(bodyParserEngineManager).getBodyParserEngineForContentType("application/json");        
-        assertTrue(o instanceof Dummy);     
-        
-        
+
+        verify(bodyParserEngineManager).getBodyParserEngineForContentType("application/json");
+        assertTrue(o instanceof Dummy);
+
+
     }
     @Test
     public void testParseBodyPostWorks() {
 
-        
+
         when(httpServletRequest.getContentType()).thenReturn(ContentTypes.APPLICATION_POST_FORM);
-        
+
         //init the context from a (mocked) servlet
         context.init(httpServletRequest, httpServletResponse);
-        
+
         when(bodyParserEngineManager.getBodyParserEngineForContentType(ContentTypes.APPLICATION_POST_FORM)).thenReturn(bodyParserEngine);
         Dummy dummy = new Dummy();
         dummy.name = "post";
         dummy.count = 245L;
         when(bodyParserEngine.invoke(context, Dummy.class)).thenReturn(dummy);
-        
+
         Dummy o = context.parseBody(Dummy.class);
-        
-        verify(bodyParserEngineManager).getBodyParserEngineForContentType(ContentTypes.APPLICATION_POST_FORM);        
-        assertTrue(o instanceof Dummy);     
-        assertTrue(o.name.equals(dummy.name));     
-        assertTrue(o.count.equals(dummy.count));     
+
+        verify(bodyParserEngineManager).getBodyParserEngineForContentType(ContentTypes.APPLICATION_POST_FORM);
+        assertTrue(o instanceof Dummy);
+        assertTrue(o.name.equals(dummy.name));
+        assertTrue(o.count.equals(dummy.count));
     }
-    
+
     /**
      * This is the default mode.
-     * 
+     *
      * We get a Content-Type: application/json and want to parse the incoming json.
      */
     @Test
     public void testParseBodyXmlWorks() {
 
-        
+
         when(httpServletRequest.getContentType()).thenReturn("application/xml");
-        
+
         //init the context from a (mocked) servlet
         context.init(httpServletRequest, httpServletResponse);
-        
+
         when(bodyParserEngineManager.getBodyParserEngineForContentType("application/xml")).thenReturn(bodyParserEngine);
         when(bodyParserEngine.invoke(context, Dummy.class)).thenReturn(new Dummy());
-        
+
         Object o = context.parseBody(Dummy.class);
-        
+
         verify(bodyParserEngineManager).getBodyParserEngineForContentType("application/xml");
-        assertTrue(o instanceof Dummy);     
-        
-        
+        assertTrue(o instanceof Dummy);
+
+
     }
-    
+
     /**
      * The request does not have the Content-Type set => we get a null response.
      */
     @Test
     public void testParseBodyWithUnkownContentTypeWorks() {
-        
+
         when(httpServletRequest.getContentType()).thenReturn(null);
-        
+
         //init the context from a (mocked) servlet
         context.init(httpServletRequest, httpServletResponse);
 
-        
+
         Object o = context.parseBody(Dummy.class);
-        
-        
-        assertNull(o);             
-        
+
+
+        assertNull(o);
+
     }
-    
+
     /**
      * We get an conetnt type that does not match any registered parsers.
      * This must also return null safely.
      */
     @Test
     public void testParseBodyWithUnknownRequestContentTypeWorks() {
-        
+
         when(httpServletRequest.getContentType()).thenReturn("application/UNKNOWN");
-        
+
         //init the context from a (mocked) servlet
         context.init(httpServletRequest, httpServletResponse);
-        
-        Object o = context.parseBody(Dummy.class);       
-        
-        assertNull(o);             
-        
+
+        Object o = context.parseBody(Dummy.class);
+
+        assertNull(o);
+
     }
-    
-    
+
+
     // Dummy class used for parseBody tests.
     class Dummy {
        public String name;
-       public Long count; 
+       public Long count;
     }
-    
-    
+
+
     /**
      * Make sure the correct character encoding is set before the
      * reader is returned.
-     * 
-     * 
+     *
+     *
      * Otherwise we might get strange default encodings
      * from the servlet engine.
-     * 
+     *
      */
     @Test
     public void testGetReaderEnforcingOfCorrectEncoding() throws Exception {
-        
+
         context.init(httpServletRequest, httpServletResponse);
-        
+
         context.getReader();
       //this proofs that the encoding has been set:
         verify(httpServletRequest).setCharacterEncoding(anyString());
-        
-        
+
+
     }
 
     /**
      * Make sure the correct character encoding is set before the
      * inputStream is returned.
-     * 
+     *
      * Otherwise we might get strange default encodings
      * from the servlet engine.
      *
      */
     @Test
     public void testGetInputStreamEnforcingOfCorrectEncoding() throws Exception {
-        
+
         context.init(httpServletRequest, httpServletResponse);
-        
+
         context.getInputStream();
         //this proofs that the encoding has been set:
         verify(httpServletRequest).setCharacterEncoding(anyString());
-        
-        
+
+
     }
-    
+
 }
